Improvements to support type-hinting
This is a big improvement for all code that utilises the
get_import_details() method on a graph object. Type-checkers can now
verify which keys are being accessed, rather than just allowing any
access on any key. They also natively understand the type of each
supported key, without us having to cast values manually.

As this project needs to support Python 3.7 for the time being, this
introduces a dependency on typing_extensions to utilise TypedDict.